refactor(utils): 重构工具函数并迁移至 composables 与组件目录
将原本位于 `src/utils` 下的逻辑根据其功能属性进行迁移和重构，以优化代码组织结构并提高模块内聚性。

- 将角色名称预设逻辑从 `utils` 迁移至 `composables/characterInfo`
- 将中文转换和 JSONL 小说转换逻辑从 `utils` 迁移至 `components/toolsbox` 目录下，使其与相关组件更紧密结合
- 优化 `imageHosting.ts` 中的图片上传逻辑，改进了 Base64 处理流程及本地资源 URL 的构建方式
- 修复 `pngCardMetadata.ts` 中的代码格式问题
- 更新相关组件及单元测试的导入路径以适配新的目录结构
